+++ src/UserDirectories.ts
@@ -4,12 +4,24 @@ import { URLBuild } from "./util/generic";
 import { IEntityRemove, IHttpStatus, ISelection } from "./types/interfaces";
 import { ITagCondensed } from "./Tags";
 import { IClassUserDirectory, UserDirectory } from "./UserDirectory";
+import { GetCommonProperties } from "./util/GetCommonProps";
+
+export type TUserDirectoryTypes =
+  | "LDAP.ActiveDirectory"
+  | "LDAP.AdvancedLDAP"
+  | "LDAP.ApacheDs"
+  | "LDAP.GenericLDAP"
+  | "ODBC.ODBC"
+  | "ODBC.OdbcAccess"
+  | "ODBC.OdbcExcel"
+  | "ODBC.OdbcSQL"
+  | "ODBC.OdbcTeradata";
 
 export interface IUserDirectoryCondensed {
   id?: string;
   privileges?: string[];
   name: string;
-  type: string;
+  type: TUserDirectoryTypes;
 }
 
 export interface IUserDirectorySettings {
@@ -19,10 +31,10 @@ export interface IUserDirectorySettings {
   modifiedByUserName?: string;
   schemaPath?: string;
   privileges?: string[];
-  name: string;
+  name?: string;
   category?: string;
-  userDirectorySettingType: number;
-  secret: boolean;
+  userDirectorySettingType?: number;
+  secret?: boolean;
   value?: string;
   secretValue?: string;
 }
@@ -43,20 +55,62 @@ export interface IUserDirectory extends IUserDirectoryCondensed {
   operationalError?: string;
   tags: ITagCondensed[];
   creationType: number;
-  settings?: IUserDirectorySettings;
+  settings?: IUserDirectorySettings[];
 }
 
 export interface IUserDirectoryUpdate {
-  id: string;
-  name: string;
-  path: string;
-  userName: string;
-  password: string;
-  ldapFilter: string;
-  timeout: string;
-  pageSize: string;
+  name?: string;
+  userDirectoryName?: string;
+  syncOnlyLoggedInUsers?: boolean;
   tags?: string[];
-  customProperties?: string[];
+  settings?: IUserDirectorySettingItem[];
+}
+
+export interface IUserDirectoryCreate {
+  name: string;
+  userDirectoryName: string;
+  type: TUserDirectoryTypes;
+  syncOnlyLoggedInUsers?: boolean;
+  tags?: string[];
+  settings?: IUserDirectorySettingItem[];
+}
+
+export interface IUserDirectorySettingItem {
+  name:
+    | "Path"
+    | "User name"
+    | "Password"
+    | "LDAP Filter"
+    | "Synchronization timeout in seconds"
+    | "Page size"
+    | "Auth type"
+    | "User directory name"
+    | "Host"
+    | "Connection timeout in seconds"
+    | "Base DN"
+    | "Use optimized query"
+    | "Flags"
+    | "Locator Flags"
+    | "Protocol Version"
+    | "Sasl Method"
+    | "Certificate path"
+    | "Attribute name of node type"
+    | "Attribute value of node type identifying a user"
+    | "Attribute value of node type identifying a group"
+    | "Attribute name of account"
+    | "Attribute name of email"
+    | "Attribute name of display name"
+    | "Attribute name of group membership"
+    | "Attribute name of members of node"
+    | "Custom Attributes"
+    | "Users table"
+    | "Attributes table"
+    | "Connection string part 1"
+    | "Connection string part 2 (secret)";
+  secret?: boolean;
+  userDirectorySettingType: "String" | "Int" | "Bool";
+  value: string;
+  secretValue?: string;
 }
 
 export interface IClassUserDirectories {
@@ -67,6 +121,7 @@ export interface IClassUserDirectories {
   removeFilter(filter: string): Promise<IEntityRemove[]>;
   select(filter?: string): Promise<ISelection>;
   syncMany(userDirectoryIds: string[]): Promise<IHttpStatus>;
+  create(arg: IUserDirectoryCreate): Promise<IClassUserDirectory>;
 }
 
 export class UserDirectories implements IClassUserDirectories {
@@ -143,5 +198,33 @@ export class UserDirectories implements IClassUserDirectories {
     return await this.repoClient
       .Post(`userdirectoryconnector/syncuserdirectories`, [...userDirectoryIds])
       .then((res) => res.status as IHttpStatus);
+  }
+
+  public async create(arg: IUserDirectoryCreate) {
+    if (!arg.name)
+      throw new Error(`userDirectories.create: "name" parameter is required`);
+    if (!arg.userDirectoryName)
+      throw new Error(
+        `userDirectories.create: "userDirectoryName" parameter is required`
+      );
+    if (!arg.type)
+      throw new Error(`userDirectories.create: "type" parameter is required`);
+
+    let obj = { ...arg } as any;
+    obj.type = `Repository.UserDirectoryConnectors.${arg.type}`;
+
+    const getCommonProps = new GetCommonProperties(
+      this.repoClient,
+      [],
+      arg.tags,
+      ""
+    );
+
+    const commonProps = await getCommonProps.getAll();
+
+    return await this.repoClient
+      .Post(`userdirectory`, { ...obj, ...commonProps })
+      .then((res) => res.data as IUserDirectory)
+      .then((ud) => new UserDirectory(this.repoClient, ud.id, ud));
   }
 }